+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Class getter and setter lens
 - Array value lens
 - Class property lens
+- Prism base class handling common behaviour